refactor(runtime): extract end_turn module from agent_loop
The end-of-turn cluster moves into `agent_loop/end_turn.rs`:

- Three context structs — `FinalizeEndTurnContext`,
  `FinalizeEndTurnResultData`, `EndTurnRetryContext` — promoted to
  `pub(super)` along with every field. The main loop in mod.rs
  literally constructs these (both `run_agent_loop` and
  `run_agent_loop_streaming` build them at four sites each), so
  literal-construction requires all fields to be visible at the
  call site.
- `EndTurnRetry` enum (`EmptyResponse { is_silent_failure }`,
  `HallucinatedAction`, `ActionIntent`) — `pub(super)`. Pattern-
  matched at two main-loop sites.
- `classify_end_turn_retry` — decides whether the just-finished
  turn should retry. Three rules: empty response (iter 0 or
  zero-token "silent failure"), hallucinated action claim with no
  actual tool call, and action-intent user message ignored by a
  text-only reply. The hallucination + action-intent branches
  delegate to `text_recovery`.
- `finalize_end_turn_text` — empty-response fallback, with the
  accumulated_text rescue for agents that emitted text alongside
  earlier `tool_use` blocks.
- `finalize_successful_end_turn` — the big success path: push
  assistant message, prune heartbeat turns, save_session_async
  (skipped on fork/incognito), proactive-memory write through
  `sanitize_for_memory`, and assemble the `AgentLoopResult`.
- `maybe_fold_stale_tool_results` — periodic history-fold to keep
  the context window viable.
- `gated_proactive_memory_for_retrieve` / `_for_memorize` —
  per-agent gates on the optional `ProactiveMemoryStore`. Both
  called from prompt setup (retrieve) and end-turn (memorize) on
  both the streaming and non-streaming paths.
- `serialize_session_messages` — stays module-private (only
  `finalize_successful_end_turn` consumes it).
- `build_silent_agent_loop_result` — bare success result for the
  silent-reply path (NO_REPLY / `[[silent]]` directives).

end_turn.rs uses `use super::*;` to inherit the parent module's
re-exports (LoopOptions, AgentLoopResult, ExperimentContext, etc.)
plus explicit `super::message::sanitize_for_memory` and
`super::text_recovery::{looks_like_hallucinated_action,
user_message_has_action_intent}` for the names not re-exported
through the wildcard.

The `mod tests` block picks up `sanitize_for_memory`,
`looks_like_hallucinated_action`, and `user_message_has_action_intent`
via the existing test-only `use super::*` pattern.

Behavior-preserving. mod.rs: 11,384 -> 10,893 LOC (-491).
cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
